Adria

See merge request pacheco.gonzalez.ivan/erp-ins-cavall-bernat!23

+++ src/main/java/cat/copernic/erpInsCavallBernat/model/Rol.java
@@ -1,10 +1,13 @@
 package cat.copernic.erpInsCavallBernat.model;
 
 import java.io.Serializable;
+import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToOne;
 import javax.persistence.Table;
 import javax.validation.constraints.NotEmpty;
 import lombok.Data;
@@ -22,8 +25,14 @@ public class Rol implements Serializable{
 
     @Id //L'atribut idRol és la clau primària de la BBDD
     @GeneratedValue(strategy=GenerationType.IDENTITY) //Generació autonumèrica de l'id
-    private int id_rol;
+    private long id_rol;
     
     @NotEmpty//Validació perquè l'usuari afegeixi contingut al camp nom
     private String nom;
+    
+//    @OneToOne(cascade = CascadeType.ALL)
+//    @JoinColumn(name = "id_usuari", referencedColumnName = "id_usuari")
+//    private Usuari id_usuari;
+ 
+    
 }
